+++ CHANGELOG.md
@@ -3,6 +3,11 @@
 <!-- next-header -->
 ## [Unreleased] (ReleaseDate)
 
+### Breaking changes
+
+* `ListSexp` now returns an `Sexp` instead of  a`TypedSexp`. Use `.into_typed()`
+  to convert an `Sexp` to a `TypedSexp`.
+
 ## [v0.2.4] (2024-01-15)
 
 ## [v0.2.2] (2024-01-15)